requirements.txt
Django~=3.2.10
djangorestframework~=3.13.1
django-rest-swagger~=2.2.0
Flask~=2.1.2
cx_Oracle
django-bootstrap-modal-forms
